Modularized the AppView a bit and formatted some code - The application was not thought out initially in the beginning and the AppView was a bit of a rough draft. I cleaned up some of the work I thought could be moved out/component-ized so that I could make PR's that don't conflict with one another, but also to make it good practice and professional. - I am pushing this work directly to 'develop' mainly because it was meant to be there, but also because it seemed pointless to create a branch and go through the process.
