+++ src/styles/App.css
@@ -39,6 +39,7 @@
 }
 .change-user-btn:hover {
   opacity: 1;
+  margin: 0 20px;
 }
 
 .App-link {
