+++ src/views/AppView.tsx
@@ -1,33 +1,23 @@
+// Packages/React
+import React from "react";
 import { ResponsiveContainer } from "recharts";
-// import allData from "../algorithm/rtnData";
+// Modularized scripts
 import spliceDataBasedOnDate from "../algorithm/spliceDataBasedOnDate";
 import initialDate from "../algorithm/initialDate";
-// import { DateTime } from "../types";
+// Components
 import Chart from "../components/Chart";
 import DatePicker from "../components/DatePicker";
 import InfoDisplay from "../components/InfoDisplay";
-import React from "react";
+import Report from "../components/Report";
+import Header from "../components/Header";
 
 function AppView() {
   // States
-  // const [currentTrans, setCurrentTrans] = React.useState({ test: "test" });
   const [name, setName] = React.useState("Camryn");
   const [dates, setDates] = React.useState(initialDate);
 
-  // Const variables that are based on data to be displayed
-  const totalExp = () => {
-    let sum: number = 0;
-    for (let trans of spliceDataBasedOnDate(dates)) {
-      sum += parseFloat(trans.CHARGE);
-    }
-    return sum.toFixed(2);
-  };
-  const returnBold = (inner: string | number) => (
-    <b className="report-bold">{inner}</b>
-  );
-
   // Const event handlers
-  const handleEvent = (event: any) => {
+  const handleNameChange = () => {
     setName(() => (name === "Camryn" ? "Kyle" : "Camryn"));
   };
   const handleDateChange = (text: string, value: string) => {
@@ -50,12 +40,7 @@ function AppView() {
 
   return (
     <div>
-      <div className="App-header">
-        <h1>WELCOME {name.toUpperCase()}!</h1>
-        <button className="change-user-btn" onClick={handleEvent}>
-          CHANGE USER
-        </button>
-      </div>
+      <Header name={name} onNameChange={handleNameChange} />
       {/* This will be the bar under the app header to display from to and anything else */}
       <div className="navbar">
         {[
@@ -81,17 +66,7 @@ function AppView() {
           <InfoDisplay data={spliceDataBasedOnDate(dates)} />
         </div>
         <div className="chart-n-rep-child">
-          <div className="report">
-            <div className="report-child">
-              Overall monthly expenses:
-              {returnBold("$" + totalExp())}
-            </div>
-            <div className="report-child">
-              Increase in spending from prev month by __%
-            </div>
-            <div className="report-child">Largest expense of the month: </div>
-            <div className="report-child">Savings: </div>
-          </div>
+          <Report dates={dates} />
         </div>
       </div>
     </div>
